% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flode.R
\docType{package}
\name{flode-package}
\alias{flode}
\alias{flode-package}
\title{flode: Forecasting Operations and Development Environment}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Flode is the Forecasting and Warning team's meta-package. Loading flode attaches all core sub-packages (reach.io, reach.hydro, reach.ensemble, reach.validate, reach.viz, reach.utils) and prints a startup message confirming what was loaded. Modelled on the tidyverse meta-package pattern.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ea-forecasting/flode}
  \item Report bugs at \url{https://github.com/ea-forecasting/flode/issues}
}

}
\author{
\strong{Maintainer}: Forecasting and Warning Team \email{forecasting@environment-agency.gov.uk}

}
\keyword{internal}
