% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flode.R
\name{flode_packages}
\alias{flode_packages}
\title{List the core flode sub-packages}
\usage{
flode_packages()
}
\value{
Character vector of sub-package names, in load order.
}
\description{
List the core flode sub-packages
}
