% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flode.R
\name{flode_update}
\alias{flode_update}
\title{Update all flode sub-packages}
\usage{
flode_update(repo = "https://git.internal/JonPayneEA", ...)
}
\arguments{
\item{repo}{Character. Base URL of the team Git platform.}

\item{...}{Additional arguments passed to \code{\link[remotes:install_git]{remotes::install_git()}}.}
}
\description{
Convenience wrapper: re-installs all flode sub-packages from the team's
internal Git repository. Requires \code{remotes}.
}
