% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{flodePackages}
\alias{flodePackages}
\title{List all packages in flode}
\usage{
flodePackages(include_self = TRUE)
}
\arguments{
\item{include_self}{Include flode in the list?}
}
\description{
List all packages in flode
}
\examples{
#flodePackages()
}
