% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{modellER_packages}
\alias{modellER_packages}
\title{List all packages in the tidyverse}
\usage{
modellER_packages(include_self = TRUE)
}
\arguments{
\item{include_self}{Include tidyverse in the list?}
}
\description{
List all packages in the tidyverse
}
\examples{
#tidyverse_packages()
}
